fix:history product table
